measures(phase-7): consolidate 10 complexity-pressure into pressure-axes + migrate script-tamper-guard + remove check-complexity-health.mjs
Per tasks.md Phase 7 sub-steps 7.1-7.6 + Luis's corrected ratification after the maintainability-budget escalation:

- New packages/measures/src/health/pressure/pressure-axes.test.ts consolidates the 10 legacy complexity-pressure axes into pressure-axes.json. It carries the script budgets verbatim, computes the legacy systems-only metrics inline, preserves PRESSURE_AXIS_REPORTS insertion order in details.axes, computes RSCD with the script formula, and writes the consolidated report through the measures report writer.
- Phase 7 consolidation discovered pre-existing dual-budget pressure vs split-test targets. The pressure rollup keeps the script's aspirational targets; Phase 4 split-test budgets stay unchanged. Duals found: cognitive 18 vs split 103, cyclomatic 20 vs split 50, maintainability 60 vs split 0, CRAP0 300 vs split 2500, boundary ratio 0.30 vs split 1, aggregate BCI 16.0 vs split 195.64, runtime fan-in 10 vs split 110. No comparable split gate was found for subdir cross-edge ratio, max file turbulence, or max package avg turbulence.
- Maintainability's dual is intentional: split-test MIN_MAINTAINABILITY_INDEX=0 is a floor/observation gate; pressure minMaintainabilityIndex=60 is an aspirational quality target. Phase 7 preserves target=60 in pressure-axes.json. Phase 12+ may converge the floor toward the target as a deliberate quality push; out of scope here.
- pressure-axes.json now fails by design with RSCD 7.614023522514476. Failing axes are min maintainability index (8.484491494694318 < 60) and max file turbulence (253 > 250). Per-axis current/budget/comparison/debtRatio/worstOffender were verified against the baseline complexity-pressure reports.
- New packages/measures/src/health/meta/script-tamper-guard.test.ts migrates guardFindings with REPO_ROOT recomputed via resolve(TEST_DIR, '../../../../..'), writes script-tamper-guard.json with a single violations rollup, and documents the Phase 12 test:codebase-health rename caveat.
- Added script-tamper-guard.test.ts to gate-integrity GATE_FILES for symmetry.
- Deleted scripts/check-complexity-health.mjs.
- Removed health:complexity, health:complexity:run, and health:complexity:report from root package.json.
- Removed ignored legacy complexity-pressure runtime JSONs; baseline copies remain audit history until the Phase-7-close snapshot lands.

After this commit, verifier runner inventory drops to vitest-only for this boundary. Gate composition shifts to report-set checks plus carry-forward and observation-only filters. A fresh Phase-7-close standardized-env snapshot should become the Phase 8+ baseline after verifier PASS.
