+++ package.json
@@ -25,9 +25,6 @@
     "health:duplication": "node --no-warnings=ExperimentalWarning --experimental-strip-types scripts/record-run.mjs --id=duplication --name=\"Code Duplication (jscpd)\" --category=Static --display=\"npm run health:duplication\" -- npm run health:duplication:run",
     "health:duplication:run": "jscpd --config .jscpd.json backend cloud_functions packages scripts tools webapp/src",
     "health:duplication:report": "jscpd --config .jscpd.json --threshold 100 backend cloud_functions packages scripts tools webapp/src",
-    "health:complexity": "node --no-warnings=ExperimentalWarning --experimental-strip-types scripts/record-run.mjs --id=complexity --name=\"Complexity Health\" --category=Static --display=\"npm run health:complexity\" -- npm run health:complexity:run",
-    "health:complexity:run": "node --no-warnings=ExperimentalWarning --experimental-strip-types scripts/check-complexity-health.mjs --report-only",
-    "health:complexity:report": "node --no-warnings=ExperimentalWarning --experimental-strip-types scripts/check-complexity-health.mjs --report-only",
     "health:mutation": "node --no-warnings=ExperimentalWarning --experimental-strip-types scripts/record-run.mjs --id=mutation --name=\"Mutation Testing (Stryker)\" --category=Other --display=\"npm run health:mutation\" --slow -- npm run health:mutation:run",
     "health:mutation:run": "npm --workspace @vt/graph-state run test:mutation:smoke",
     "lint": "node --no-warnings=ExperimentalWarning --experimental-strip-types scripts/record-run.mjs --id=root-lint --name=\"Root ESLint\" --category=Lint --display=\"npm run lint\" -- npm run lint:run",
