+++ packages/measures/src/health/meta/script-tamper-guard.test.ts
@@ -1,0 +1,117 @@
+import {execSync} from 'node:child_process'
+import {existsSync, readFileSync} from 'node:fs'
+import {basename, dirname, join, resolve} from 'node:path'
+import {fileURLToPath} from 'node:url'
+import {describe, expect, it} from 'vitest'
+import {recordHealthMetric} from '../../_shared/report-writer'
+
+const TEST_DIR: string = dirname(fileURLToPath(import.meta.url))
+const REPO_ROOT: string = resolve(TEST_DIR, '../../../../..')
+
+function runGit(args: string): string {
+    return execSync(`git ${args}`, {cwd: REPO_ROOT, encoding: 'utf8', stdio: ['pipe', 'pipe', 'pipe']})
+}
+
+function tryRunGit(args: string): string | null {
+    try {
+        return runGit(args)
+    } catch {
+        return null
+    }
+}
+
+function changedStatusEntries(): string[] {
+    const output = tryRunGit('status --porcelain') ?? ''
+    return output.split('\n').map(line => line.trimEnd()).filter(Boolean)
+}
+
+function isExpectedCodebaseHealthScriptMove(scriptName: string, current: unknown, committed: unknown): boolean {
+    const legacyPackage = '@vt/' + 'codebase-health'
+    // Phase 12 will rename test:codebase-health to test:measures; update this
+    // heuristic in that phase rather than widening it during Phase 7.
+    return scriptName === 'test:codebase-health'
+        && current === 'npm --workspace @vt/measures run test'
+        && typeof committed === 'string'
+        && (
+            committed === `npm --workspace ${legacyPackage} run test`
+            || committed.startsWith('vitest run packages/systems/')
+        )
+}
+
+function isMovedCodebaseHealthTest(path: string): boolean {
+    if (!/^packages\/codebase-health\/src\/[^/]+\.test\.ts$/.test(path)) return false
+    const movedHealthDirs = ['churn', 'complexity', 'coupling', 'meta', 'purity', 'shape']
+    return movedHealthDirs.some(dir => existsSync(join(REPO_ROOT, 'packages/measures/src/health', dir, basename(path))))
+}
+
+function packageJsonScriptFindings(): string[] {
+    const findings: string[] = []
+    const currentPackageJson = JSON.parse(readFileSync(join(REPO_ROOT, 'package.json'), 'utf8')) as {scripts?: Record<string, string>}
+    const committedPackageJsonRaw = tryRunGit('show HEAD:package.json')
+    if (!committedPackageJsonRaw) return findings
+
+    const committedPackageJson = JSON.parse(committedPackageJsonRaw) as {scripts?: Record<string, string>}
+    for (const scriptName of ['test', 'test:codebase-health']) {
+        const currentScript = currentPackageJson.scripts?.[scriptName]
+        const committedScript = committedPackageJson.scripts?.[scriptName]
+        if (
+            currentScript !== committedScript
+            && !isExpectedCodebaseHealthScriptMove(scriptName, currentScript, committedScript)
+        ) {
+            findings.push(`package.json script "${scriptName}" changed; complexity pressure must not be won by relaxing green gates`)
+        }
+    }
+
+    return findings
+}
+
+function deletedTestFindings(): string[] {
+    const deletedTests = changedStatusEntries()
+        .filter(line => line.startsWith('D ') || line.startsWith(' D'))
+        .map(line => line.slice(2).trim())
+        .filter(path => /\.(test|spec)\.(ts|tsx|js|jsx)$/.test(path))
+        .filter(path => !isMovedCodebaseHealthTest(path))
+
+    return deletedTests.length === 0 ? [] : [`deleted test files detected: ${deletedTests.join(', ')}`]
+}
+
+function missingGateFileFindings(): string[] {
+    const requiredFiles = [
+        'packages/measures/src/health/meta/gate-integrity.test.ts',
+        'packages/measures/src/health/meta/script-tamper-guard.test.ts',
+        'packages/measures/src/health/purity/purity-ratio-ast.test.ts',
+        'packages/measures/src/health/complexity/cognitive-complexity.test.ts',
+        'packages/measures/src/health/coupling/cross-package-coupling.test.ts',
+    ]
+
+    return requiredFiles
+        .filter(file => !existsSync(join(REPO_ROOT, file)))
+        .map(file => `required health gate file missing: ${file}`)
+}
+
+function guardFindings(): string[] {
+    return [
+        ...packageJsonScriptFindings(),
+        ...deletedTestFindings(),
+        ...missingGateFileFindings(),
+    ]
+}
+
+describe('script tamper guard', () => {
+    it('detects unauthorized modifications to npm scripts and gate tests', async () => {
+        const violations = guardFindings()
+
+        await recordHealthMetric({
+            metricId: 'script-tamper-guard',
+            metricName: 'Script Tamper Guard',
+            description: 'Detects unauthorized modifications to npm scripts',
+            category: 'Other',
+            current: violations.length,
+            budget: 0,
+            comparison: 'lte',
+            details: {violations},
+        })
+
+        expect(violations, violations.join('\n')).toEqual([])
+    })
+})
