+++ packages/measures/src/health/pressure/pressure-axes.test.ts
@@ -1,0 +1,891 @@
+import {execSync} from 'node:child_process'
+import {readdir, readFile, stat} from 'node:fs/promises'
+import {dirname, join, relative, resolve} from 'node:path'
+import {fileURLToPath} from 'node:url'
+import * as ts from 'typescript'
+import {describe, expect, it} from 'vitest'
+import {recordHealthMetric} from '../../_shared/report-writer'
+
+const TEST_DIR = dirname(fileURLToPath(import.meta.url))
+const REPO_ROOT = resolve(TEST_DIR, '../../../../..')
+const SYSTEMS_ROOT = join(REPO_ROOT, 'packages', 'systems')
+const REPORTS_DIR = join(REPO_ROOT, 'health-dashboard', 'reports')
+
+const PRESSURE_AXIS_CONFIGS = [
+    // Aspirational function readability target from the legacy pressure rollup.
+    {
+        name: 'max cognitive complexity',
+        metricKey: 'maxCognitiveComplexity',
+        metricId: 'complexity-pressure-cognitive-max',
+        budget: 18,
+        comparison: 'lte',
+        unit: 'score',
+    },
+    // Stricter systems-only pressure target than the whole-repo split test.
+    {
+        name: 'max cyclomatic complexity',
+        metricKey: 'maxCyclomaticComplexity',
+        metricId: 'complexity-pressure-cyclomatic-max',
+        budget: 20,
+        comparison: 'lte',
+        unit: 'score',
+    },
+    // Aspirational MI quality target; distinct from the Phase 4 floor gate.
+    {
+        name: 'min maintainability index',
+        metricKey: 'minMaintainabilityIndex',
+        metricId: 'complexity-pressure-maintainability-min',
+        budget: 60,
+        comparison: 'gte',
+        unit: 'index',
+    },
+    // Stricter systems-only zero-coverage risk target than the whole-repo split test.
+    {
+        name: 'max CRAP0 risk',
+        metricKey: 'maxCrapZeroCoverage',
+        metricId: 'complexity-pressure-crap0-max',
+        budget: 300,
+        comparison: 'lte',
+        unit: 'score',
+    },
+    // Systems graph target for files touching package boundaries.
+    {
+        name: 'max boundary ratio',
+        metricKey: 'maxBoundaryRatio',
+        metricId: 'complexity-pressure-boundary-ratio-max',
+        budget: 0.30,
+        comparison: 'lte',
+        unit: 'ratio',
+    },
+    // Systems graph target for intra-package imports that cross source subdirectories.
+    {
+        name: 'max subdirectory cross-edge ratio',
+        metricKey: 'maxSubdirCrossRatio',
+        metricId: 'complexity-pressure-subdir-cross-ratio-max',
+        budget: 0.60,
+        comparison: 'lte',
+        unit: 'ratio',
+    },
+    // Systems graph aggregate BCI target carried from the script.
+    {
+        name: 'aggregate boundary complexity',
+        metricKey: 'aggregateBoundaryComplexity',
+        metricId: 'complexity-pressure-boundary-complexity-aggregate',
+        budget: 16.0,
+        comparison: 'lte',
+        unit: 'bci',
+    },
+    // Systems-only runtime fan-in target from the legacy pressure rollup.
+    {
+        name: 'max runtime fan-in',
+        metricKey: 'maxRuntimeFanIn',
+        metricId: 'complexity-pressure-runtime-fan-in-max',
+        budget: 10,
+        comparison: 'lte',
+        unit: 'symbols',
+    },
+    // Churn multiplied by structural complexity target for individual systems files.
+    {
+        name: 'max file turbulence',
+        metricKey: 'maxFileTurbulence',
+        metricId: 'complexity-pressure-file-turbulence-max',
+        budget: 250,
+        comparison: 'lte',
+        unit: 'turbulence',
+    },
+    // Churn multiplied by structural complexity target for systems package averages.
+    {
+        name: 'max package avg turbulence',
+        metricKey: 'maxPackageAverageTurbulence',
+        metricId: 'complexity-pressure-package-turbulence-avg-max',
+        budget: 35,
+        comparison: 'lte',
+        unit: 'turbulence',
+    },
+] as const
+
+type PressureAxisConfig = typeof PRESSURE_AXIS_CONFIGS[number]
+
+type SystemPackage = {
+    readonly name: string
+    readonly dirName: string
+    readonly srcRoot: string
+}
+
+type SystemFile = {
+    readonly absolutePath: string
+    readonly relativePath: string
+    readonly packageName: string
+    readonly npmName: string
+    readonly subdirectory: string
+}
+
+type GraphEdge = {
+    readonly from: string
+    readonly to: string
+    readonly fromPackage: string
+    readonly toPackage: string
+    readonly fromSubdirectory: string
+    readonly toSubdirectory: string
+}
+
+type SystemGraph = {
+    readonly files: readonly SystemFile[]
+    readonly edges: readonly GraphEdge[]
+    readonly runtimeSymbolsByTarget: ReadonlyMap<string, ReadonlyMap<string, ReadonlySet<string>>>
+}
+
+type FunctionComplexity = {
+    readonly packageName: string
+    readonly file: string
+    readonly line: number
+    readonly name: string
+    readonly score: number
+    readonly crapZeroCoverage?: number
+}
+
+type MaintainabilityRow = {
+    readonly file: string
+    readonly maintainabilityIndex: number
+}
+
+type TurbulenceRow = {
+    readonly packageName: string
+    readonly file: string
+    readonly churn: number
+    readonly complexity: number
+    readonly turbulence: number
+}
+
+type PressureAxis = {
+    readonly name: string
+    readonly metricKey: PressureAxisConfig['metricKey']
+    readonly current: number
+    readonly budget: number
+    readonly comparison: 'lte' | 'gte'
+    readonly passed: boolean
+    readonly debtRatio: number
+    readonly worstOffender: string
+}
+
+type LegacyPressureReport = {
+    readonly metricId: string
+    readonly current: number
+    readonly budget: number
+    readonly comparison: 'lte' | 'gte'
+    readonly details?: {
+        readonly debtRatio?: number
+        readonly worstOffender?: string
+    }
+}
+
+async function pathExists(path: string): Promise<boolean> {
+    try {
+        await stat(path)
+        return true
+    } catch {
+        return false
+    }
+}
+
+async function discoverSystemPackages(): Promise<SystemPackage[]> {
+    const entries = await readdir(SYSTEMS_ROOT, {withFileTypes: true})
+    const packages = await Promise.all(entries.map(async entry => {
+        if (!entry.isDirectory()) return null
+        const packageJsonPath = join(SYSTEMS_ROOT, entry.name, 'package.json')
+        const srcRoot = join(SYSTEMS_ROOT, entry.name, 'src')
+        if (!(await pathExists(packageJsonPath)) || !(await pathExists(srcRoot))) return null
+        const packageJson = JSON.parse(await readFile(packageJsonPath, 'utf8')) as {name?: string}
+        if (!packageJson.name) return null
+        return {name: packageJson.name, dirName: entry.name, srcRoot}
+    }))
+    return packages.filter((pkg): pkg is SystemPackage => pkg !== null).sort((a, b) => a.dirName.localeCompare(b.dirName))
+}
+
+function isProductionSource(path: string): boolean {
+    return path.endsWith('.ts')
+        && !path.endsWith('.test.ts')
+        && !path.endsWith('.spec.ts')
+        && !path.endsWith('.d.ts')
+        && !path.includes('/__tests__/')
+}
+
+async function listProductionSources(root: string): Promise<string[]> {
+    if (!(await pathExists(root))) return []
+    const entries = await readdir(root, {withFileTypes: true})
+    const nested = await Promise.all(entries.map(async entry => {
+        const path = join(root, entry.name)
+        if (entry.isDirectory()) return listProductionSources(path)
+        if (entry.isFile() && isProductionSource(path)) return [path]
+        return []
+    }))
+    return nested.flat().sort()
+}
+
+function subdirectoryOf(absolutePath: string, srcRoot: string): string {
+    const srcRelative = relative(srcRoot, absolutePath)
+    const firstSlash = srcRelative.indexOf('/')
+    return firstSlash >= 0 ? srcRelative.slice(0, firstSlash) : '.'
+}
+
+function extractImportDeclarations(filePath: string, text: string): {specifier: string; isTypeOnly: boolean; text: string}[] {
+    const sourceFile = ts.createSourceFile(filePath, text, ts.ScriptTarget.Latest, true)
+    const declarations: {specifier: string; isTypeOnly: boolean; text: string}[] = []
+
+    for (const statement of sourceFile.statements) {
+        if (ts.isImportDeclaration(statement) && ts.isStringLiteral(statement.moduleSpecifier)) {
+            declarations.push({
+                specifier: statement.moduleSpecifier.text,
+                isTypeOnly: statement.importClause?.isTypeOnly ?? false,
+                text: statement.getText(sourceFile),
+            })
+        }
+        if (ts.isExportDeclaration(statement) && statement.moduleSpecifier && ts.isStringLiteral(statement.moduleSpecifier)) {
+            declarations.push({
+                specifier: statement.moduleSpecifier.text,
+                isTypeOnly: statement.isTypeOnly,
+                text: statement.getText(sourceFile),
+            })
+        }
+    }
+
+    return declarations
+}
+
+function resolveFileCandidate(basePath: string, knownFiles: ReadonlySet<string>): string | null {
+    const resolved = resolve(basePath)
+    const candidates = resolved.endsWith('.ts')
+        ? [resolved]
+        : [resolved, `${resolved}.ts`, join(resolved, 'index.ts')]
+    return candidates.find(candidate => knownFiles.has(candidate)) ?? null
+}
+
+function resolveSpecifier(
+    fromAbsPath: string,
+    specifier: string,
+    packagesByNpmName: ReadonlyMap<string, SystemPackage>,
+    knownFiles: ReadonlySet<string>,
+): string | null {
+    if (specifier.startsWith('.')) return resolveFileCandidate(join(dirname(fromAbsPath), specifier), knownFiles)
+
+    for (const [npmName, pkg] of packagesByNpmName) {
+        if (specifier !== npmName && !specifier.startsWith(`${npmName}/`)) continue
+        const subPath = specifier === npmName ? 'index' : specifier.slice(npmName.length + 1)
+        return resolveFileCandidate(join(pkg.srcRoot, subPath), knownFiles)
+    }
+
+    return null
+}
+
+function collectRuntimeSymbols(declaration: {isTypeOnly: boolean; text: string}): string[] {
+    if (declaration.isTypeOnly) return []
+    const match = declaration.text.match(/(?:import|export)\s*(?:type\s*)?\{([^}]*)\}/)
+    if (!match) return declaration.text.includes('* as ') ? ['*'] : []
+
+    return match[1]
+        .split(',')
+        .map(part => part.trim())
+        .filter(Boolean)
+        .filter(part => !part.startsWith('type '))
+        .map(part => part.split(/\s+as\s+/)[0].trim())
+        .filter(Boolean)
+}
+
+async function buildSystemGraph(packages: readonly SystemPackage[]): Promise<SystemGraph> {
+    const files: SystemFile[] = []
+    for (const pkg of packages) {
+        const paths = await listProductionSources(pkg.srcRoot)
+        files.push(...paths.map(absolutePath => ({
+            absolutePath: resolve(absolutePath),
+            relativePath: relative(REPO_ROOT, absolutePath),
+            packageName: pkg.dirName,
+            npmName: pkg.name,
+            subdirectory: subdirectoryOf(resolve(absolutePath), pkg.srcRoot),
+        })))
+    }
+
+    const filesByPath = new Map(files.map(file => [file.absolutePath, file]))
+    const knownFiles = new Set(filesByPath.keys())
+    const packagesByNpmName = new Map(packages.map(pkg => [pkg.name, pkg]))
+    const edges: GraphEdge[] = []
+    const runtimeSymbolsByTarget = new Map<string, Map<string, Set<string>>>()
+    const seenFileEdges = new Set<string>()
+
+    for (const fromFile of files) {
+        const text = await readFile(fromFile.absolutePath, 'utf8')
+        for (const declaration of extractImportDeclarations(fromFile.absolutePath, text)) {
+            const toPath = resolveSpecifier(fromFile.absolutePath, declaration.specifier, packagesByNpmName, knownFiles)
+            const toFile = toPath ? filesByPath.get(toPath) : null
+            if (toFile && toFile.absolutePath !== fromFile.absolutePath) {
+                const edgeKey = `${fromFile.relativePath}\0${toFile.relativePath}`
+                if (!seenFileEdges.has(edgeKey)) {
+                    seenFileEdges.add(edgeKey)
+                    edges.push({
+                        from: fromFile.relativePath,
+                        to: toFile.relativePath,
+                        fromPackage: fromFile.packageName,
+                        toPackage: toFile.packageName,
+                        fromSubdirectory: fromFile.subdirectory,
+                        toSubdirectory: toFile.subdirectory,
+                    })
+                }
+            }
+
+            const targetPkg = packages.find(pkg => declaration.specifier === pkg.name || declaration.specifier.startsWith(`${pkg.name}/`))
+            if (!targetPkg || targetPkg.dirName === fromFile.packageName) continue
+            const targetSymbols = runtimeSymbolsByTarget.get(targetPkg.dirName) ?? new Map<string, Set<string>>()
+            runtimeSymbolsByTarget.set(targetPkg.dirName, targetSymbols)
+            for (const symbol of collectRuntimeSymbols(declaration)) {
+                const filesForSymbol = targetSymbols.get(symbol) ?? new Set<string>()
+                filesForSymbol.add(fromFile.relativePath)
+                targetSymbols.set(symbol, filesForSymbol)
+            }
+        }
+    }
+
+    return {files, edges, runtimeSymbolsByTarget}
+}
+
+function isLogicalOperator(kind: ts.SyntaxKind): boolean {
+    return kind === ts.SyntaxKind.AmpersandAmpersandToken
+        || kind === ts.SyntaxKind.BarBarToken
+        || kind === ts.SyntaxKind.QuestionQuestionToken
+}
+
+function isLogicalExpression(node: ts.Node): node is ts.BinaryExpression {
+    return ts.isBinaryExpression(node) && isLogicalOperator(node.operatorToken.kind)
+}
+
+function countLogicalOperatorChains(expression: ts.Expression): number {
+    const operators: ts.SyntaxKind[] = []
+    function collect(node: ts.Node): void {
+        if (!isLogicalExpression(node)) return
+        collect(node.left)
+        operators.push(node.operatorToken.kind)
+        collect(node.right)
+    }
+    collect(expression)
+    if (operators.length === 0) return 0
+    let chains = 1
+    for (let i = 1; i < operators.length; i += 1) {
+        if (operators[i] !== operators[i - 1]) chains += 1
+    }
+    return chains
+}
+
+function propertyNameText(name: ts.PropertyName, sourceFile: ts.SourceFile): string {
+    if (ts.isIdentifier(name) || ts.isStringLiteral(name) || ts.isNumericLiteral(name)) return name.text
+    return name.getText(sourceFile)
+}
+
+function functionName(node: ts.Node, sourceFile: ts.SourceFile): string {
+    if ((ts.isFunctionDeclaration(node) || ts.isFunctionExpression(node)) && node.name) return node.name.text
+    if (ts.isMethodDeclaration(node) && node.name) return propertyNameText(node.name, sourceFile)
+    if (ts.isConstructorDeclaration(node)) return 'constructor'
+    if ((ts.isArrowFunction(node) || ts.isFunctionExpression(node)) && ts.isVariableDeclaration(node.parent) && ts.isIdentifier(node.parent.name)) return node.parent.name.text
+    if ((ts.isArrowFunction(node) || ts.isFunctionExpression(node)) && ts.isPropertyAssignment(node.parent)) return propertyNameText(node.parent.name, sourceFile)
+    return '<anonymous>'
+}
+
+function isFunctionLikeBoundary(node: ts.Node): node is ts.FunctionLikeDeclaration {
+    return ts.isFunctionDeclaration(node)
+        || ts.isFunctionExpression(node)
+        || ts.isArrowFunction(node)
+        || ts.isMethodDeclaration(node)
+        || ts.isConstructorDeclaration(node)
+}
+
+function isDirectRecursiveCall(node: ts.CallExpression, name: string): boolean {
+    if (name === '<anonymous>' || name === 'constructor') return false
+    if (ts.isIdentifier(node.expression)) return node.expression.text === name
+    if (ts.isPropertyAccessExpression(node.expression)) return node.expression.name.text === name
+    return false
+}
+
+function scoreFunction(root: ts.FunctionLikeDeclaration, name: string, sourceFile: ts.SourceFile): number {
+    let score = 0
+    const addStructural = (nesting: number): void => { score += 1 + nesting }
+
+    function visitIfStatement(node: ts.IfStatement, nesting: number, isElseIf: boolean): void {
+        if (isElseIf) score += 1
+        else addStructural(nesting)
+        visit(node.expression, nesting)
+        visit(node.thenStatement, nesting + 1)
+        if (!node.elseStatement) return
+        if (ts.isIfStatement(node.elseStatement)) {
+            visitIfStatement(node.elseStatement, nesting, true)
+            return
+        }
+        score += 1
+        visit(node.elseStatement, nesting + 1)
+    }
+
+    function visit(node: ts.Node, nesting: number): void {
+        if (node !== root && isFunctionLikeBoundary(node)) return
+        if (ts.isIfStatement(node)) {
+            visitIfStatement(node, nesting, false)
+            return
+        }
+        if (ts.isForStatement(node) || ts.isForInStatement(node) || ts.isForOfStatement(node) || ts.isWhileStatement(node) || ts.isDoStatement(node)) {
+            addStructural(nesting)
+            ts.forEachChild(node, child => visit(child, nesting + 1))
+            return
+        }
+        if (ts.isSwitchStatement(node)) {
+            for (const clause of node.caseBlock.clauses) {
+                if (ts.isCaseClause(clause)) score += 1 + nesting
+                ts.forEachChild(clause, child => visit(child, nesting + 1))
+            }
+            return
+        }
+        if (ts.isCatchClause(node)) {
+            addStructural(nesting)
+            visit(node.block, nesting + 1)
+            return
+        }
+        if (ts.isConditionalExpression(node)) {
+            addStructural(nesting)
+            ts.forEachChild(node, child => visit(child, nesting + 1))
+            return
+        }
+        if ((ts.isBreakStatement(node) || ts.isContinueStatement(node)) && node.label) score += 1
+        if (ts.isCallExpression(node) && isDirectRecursiveCall(node, name)) score += 1
+        if (isLogicalExpression(node) && !isLogicalExpression(node.parent)) score += countLogicalOperatorChains(node)
+        ts.forEachChild(node, child => visit(child, nesting))
+    }
+
+    visit(root, 0)
+    return score
+}
+
+async function measureCognitiveComplexity(files: readonly SystemFile[]): Promise<FunctionComplexity[]> {
+    const rows: FunctionComplexity[] = []
+    for (const file of files) {
+        const text = await readFile(file.absolutePath, 'utf8')
+        const sourceFile = ts.createSourceFile(file.absolutePath, text, ts.ScriptTarget.Latest, true)
+        function visit(node: ts.Node): void {
+            if (isFunctionLikeBoundary(node)) {
+                const name = functionName(node, sourceFile)
+                const {line} = sourceFile.getLineAndCharacterOfPosition(node.getStart(sourceFile))
+                rows.push({
+                    packageName: file.packageName,
+                    file: file.relativePath,
+                    line: line + 1,
+                    name,
+                    score: scoreFunction(node, name, sourceFile),
+                })
+            }
+            ts.forEachChild(node, visit)
+        }
+        ts.forEachChild(sourceFile, visit)
+    }
+    return rows.sort((a, b) => b.score - a.score || a.file.localeCompare(b.file))
+}
+
+function cyclomaticIncrement(node: ts.Node): number {
+    if (ts.isIfStatement(node)
+        || ts.isForStatement(node)
+        || ts.isForInStatement(node)
+        || ts.isForOfStatement(node)
+        || ts.isWhileStatement(node)
+        || ts.isDoStatement(node)
+        || ts.isCatchClause(node)
+        || ts.isConditionalExpression(node)) {
+        return 1
+    }
+    if (ts.isCaseClause(node)) return 1
+    if (isLogicalExpression(node)) return 1
+    return 0
+}
+
+function scoreCyclomaticComplexity(root: ts.FunctionLikeDeclaration): number {
+    let score = 1
+    function visit(node: ts.Node): void {
+        if (node !== root && isFunctionLikeBoundary(node)) return
+        score += cyclomaticIncrement(node)
+        ts.forEachChild(node, visit)
+    }
+    visit(root)
+    return score
+}
+
+async function measureCyclomaticComplexity(files: readonly SystemFile[]): Promise<Required<FunctionComplexity>[]> {
+    const rows: Required<FunctionComplexity>[] = []
+    for (const file of files) {
+        const text = await readFile(file.absolutePath, 'utf8')
+        const sourceFile = ts.createSourceFile(file.absolutePath, text, ts.ScriptTarget.Latest, true)
+        function visit(node: ts.Node): void {
+            if (isFunctionLikeBoundary(node)) {
+                const name = functionName(node, sourceFile)
+                const {line} = sourceFile.getLineAndCharacterOfPosition(node.getStart(sourceFile))
+                const score = scoreCyclomaticComplexity(node)
+                rows.push({
+                    packageName: file.packageName,
+                    file: file.relativePath,
+                    line: line + 1,
+                    name,
+                    score,
+                    crapZeroCoverage: score * score + score,
+                })
+            }
+            ts.forEachChild(node, visit)
+        }
+        ts.forEachChild(sourceFile, visit)
+    }
+    return rows.sort((a, b) => b.score - a.score || a.file.localeCompare(b.file))
+}
+
+function sourceLinesOfCode(text: string): number {
+    return text.split('\n')
+        .map(line => line.trim())
+        .filter(line => line && !line.startsWith('//') && !line.startsWith('*') && !line.startsWith('/*'))
+        .length
+}
+
+function isOperatorToken(kind: ts.SyntaxKind): boolean {
+    return (kind >= ts.SyntaxKind.FirstKeyword && kind <= ts.SyntaxKind.LastKeyword)
+        || (kind >= ts.SyntaxKind.FirstPunctuation && kind <= ts.SyntaxKind.LastPunctuation)
+}
+
+function isOperandToken(kind: ts.SyntaxKind): boolean {
+    return kind === ts.SyntaxKind.Identifier
+        || kind === ts.SyntaxKind.PrivateIdentifier
+        || kind === ts.SyntaxKind.NumericLiteral
+        || kind === ts.SyntaxKind.BigIntLiteral
+        || kind === ts.SyntaxKind.StringLiteral
+        || kind === ts.SyntaxKind.RegularExpressionLiteral
+        || kind === ts.SyntaxKind.NoSubstitutionTemplateLiteral
+}
+
+function measureHalstead(filePath: string, text: string, cyclomatic: number): MaintainabilityRow {
+    const scanner = ts.createScanner(ts.ScriptTarget.Latest, false, ts.LanguageVariant.Standard, text)
+    const operators = new Map<string, number>()
+    const operands = new Map<string, number>()
+    let token = scanner.scan()
+
+    while (token !== ts.SyntaxKind.EndOfFileToken) {
+        const value = scanner.getTokenText()
+        if (isOperatorToken(token)) operators.set(value, (operators.get(value) ?? 0) + 1)
+        if (isOperandToken(token)) operands.set(value, (operands.get(value) ?? 0) + 1)
+        token = scanner.scan()
+    }
+
+    const n1 = operators.size
+    const n2 = operands.size
+    const totalOperators = [...operators.values()].reduce((sum, count) => sum + count, 0)
+    const totalOperands = [...operands.values()].reduce((sum, count) => sum + count, 0)
+    const vocabulary = n1 + n2
+    const length = totalOperators + totalOperands
+    const volume = vocabulary === 0 || length === 0 ? 0 : length * Math.log2(vocabulary)
+    const sloc = sourceLinesOfCode(text)
+    const rawMaintainability = 171
+        - 5.2 * Math.log(Math.max(1, volume))
+        - 0.23 * cyclomatic
+        - 16.2 * Math.log(Math.max(1, sloc))
+    const maintainabilityIndex = Math.max(0, Math.min(100, (rawMaintainability * 100) / 171))
+
+    return {file: relative(REPO_ROOT, filePath), maintainabilityIndex}
+}
+
+async function measureMaintainability(files: readonly SystemFile[], cyclomaticRows: readonly FunctionComplexity[]): Promise<MaintainabilityRow[]> {
+    const cyclomaticByFile = new Map<string, number>()
+    for (const row of cyclomaticRows) {
+        cyclomaticByFile.set(row.file, (cyclomaticByFile.get(row.file) ?? 0) + row.score)
+    }
+
+    const rows: MaintainabilityRow[] = []
+    for (const file of files) {
+        const text = await readFile(file.absolutePath, 'utf8')
+        rows.push(measureHalstead(file.absolutePath, text, cyclomaticByFile.get(file.relativePath) ?? 1))
+    }
+    return rows.sort((a, b) => a.maintainabilityIndex - b.maintainabilityIndex || a.file.localeCompare(b.file))
+}
+
+function countSimpleComplexity(filePath: string, text: string): number {
+    const sourceFile = ts.createSourceFile(filePath, text, ts.ScriptTarget.Latest, true)
+    let complexity = 0
+    function visit(node: ts.Node): void {
+        if (ts.isIfStatement(node)
+            || ts.isForStatement(node)
+            || ts.isForInStatement(node)
+            || ts.isForOfStatement(node)
+            || ts.isWhileStatement(node)
+            || ts.isDoStatement(node)
+            || ts.isSwitchStatement(node)
+            || ts.isCatchClause(node)
+            || ts.isConditionalExpression(node)) {
+            complexity += 1
+        }
+        ts.forEachChild(node, visit)
+    }
+    ts.forEachChild(sourceFile, visit)
+    return complexity
+}
+
+function tryRunGit(args: string): string | null {
+    try {
+        return execSync(`git ${args}`, {cwd: REPO_ROOT, encoding: 'utf8', stdio: ['pipe', 'pipe', 'pipe']})
+    } catch {
+        return null
+    }
+}
+
+function collectGitChurn(): ReadonlyMap<string, number> {
+    const output = tryRunGit("log --since='6 months ago' --format=%H --name-only -- packages/systems") ?? ''
+    const churn = new Map<string, number>()
+    for (const line of output.split('\n')) {
+        const file = line.trim()
+        if (!file || !file.startsWith('packages/systems/')) continue
+        churn.set(file, (churn.get(file) ?? 0) + 1)
+    }
+    return churn
+}
+
+async function measureTurbulence(files: readonly SystemFile[]): Promise<TurbulenceRow[]> {
+    const churn = collectGitChurn()
+    const rows: TurbulenceRow[] = []
+    for (const file of files) {
+        const text = await readFile(file.absolutePath, 'utf8')
+        const fileChurn = churn.get(file.relativePath) ?? 0
+        const complexity = countSimpleComplexity(file.absolutePath, text)
+        rows.push({
+            packageName: file.packageName,
+            file: file.relativePath,
+            churn: fileChurn,
+            complexity,
+            turbulence: fileChurn * complexity,
+        })
+    }
+    return rows.sort((a, b) => b.turbulence - a.turbulence || a.file.localeCompare(b.file))
+}
+
+function mcsTreeWidthLowerBound(nodes: readonly string[], pairs: readonly (readonly [string, string])[]): number {
+    if (nodes.length <= 1) return 0
+    const adjacency = new Map(nodes.map(node => [node, new Set<string>()]))
+    for (const [a, b] of pairs) {
+        adjacency.get(a)?.add(b)
+        adjacency.get(b)?.add(a)
+    }
+
+    const numbered = new Set<string>()
+    let maxWidth = 0
+    for (let i = 0; i < nodes.length; i += 1) {
+        let bestNode = ''
+        let bestCount = -1
+        for (const node of nodes) {
+            if (numbered.has(node)) continue
+            let count = 0
+            for (const neighbor of adjacency.get(node) ?? []) {
+                if (numbered.has(neighbor)) count += 1
+            }
+            if (count > bestCount) {
+                bestNode = node
+                bestCount = count
+            }
+        }
+        if (bestCount > 0) maxWidth = Math.max(maxWidth, bestCount)
+        numbered.add(bestNode)
+    }
+    return maxWidth
+}
+
+function measureBoundaries(files: readonly SystemFile[], edges: readonly GraphEdge[], packageNames: readonly string[]) {
+    const boundaryFiles = new Map(packageNames.map(name => [name, new Set<string>()]))
+    for (const edge of edges) {
+        if (edge.fromPackage === edge.toPackage) continue
+        boundaryFiles.get(edge.fromPackage)?.add(edge.from)
+        boundaryFiles.get(edge.toPackage)?.add(edge.to)
+    }
+
+    const filesByPackage = new Map(packageNames.map(name => [name, files.filter(file => file.packageName === name)]))
+    const boundaryProfiles = packageNames.map(packageName => {
+        const totalFiles = filesByPackage.get(packageName)?.length ?? 0
+        const count = boundaryFiles.get(packageName)?.size ?? 0
+        return {packageName, boundaryFiles: count, totalFiles, ratio: totalFiles === 0 ? 0 : count / totalFiles}
+    }).sort((a, b) => b.ratio - a.ratio)
+
+    const subdirProfiles = packageNames.map(packageName => {
+        const internalEdges = edges.filter(edge => edge.fromPackage === packageName && edge.toPackage === packageName)
+        const crossSubdirEdges = internalEdges.filter(edge => edge.fromSubdirectory !== edge.toSubdirectory)
+        return {packageName, internalEdges: internalEdges.length, crossSubdirEdges: crossSubdirEdges.length, ratio: internalEdges.length === 0 ? 0 : crossSubdirEdges.length / internalEdges.length}
+    }).sort((a, b) => b.ratio - a.ratio)
+
+    const pairGroups = new Map<string, GraphEdge[]>()
+    for (const edge of edges) {
+        if (edge.fromPackage === edge.toPackage) continue
+        const key = `${edge.fromPackage} -> ${edge.toPackage}`
+        const pairEdges = pairGroups.get(key) ?? []
+        pairEdges.push(edge)
+        pairGroups.set(key, pairEdges)
+    }
+
+    const pairMetrics = [...pairGroups.entries()].map(([pair, pairEdges]) => {
+        const src = new Set(pairEdges.map(edge => edge.from))
+        const tgt = new Set(pairEdges.map(edge => edge.to))
+        const srcNodes = [...src].map(file => `src:${file}`)
+        const tgtNodes = [...tgt].map(file => `tgt:${file}`)
+        const pairs = pairEdges.map(edge => [`src:${edge.from}`, `tgt:${edge.to}`] as const)
+        const treeWidth = mcsTreeWidthLowerBound([...srcNodes, ...tgtNodes], pairs)
+        const density = src.size === 0 || tgt.size === 0 ? 0 : pairEdges.length / (src.size * tgt.size)
+        return {pair, srcFan: src.size, tgtFan: tgt.size, edgeCount: pairEdges.length, density, treeWidth, bci: (treeWidth + 1) * Math.log2(pairEdges.length + 1)}
+    }).sort((a, b) => b.bci - a.bci || a.pair.localeCompare(b.pair))
+
+    return {
+        boundaryProfiles,
+        subdirProfiles,
+        pairMetrics,
+        aggregateBci: pairMetrics.reduce((sum, pair) => sum + pair.bci, 0),
+    }
+}
+
+function runtimeFanInRows(runtimeSymbolsByTarget: SystemGraph['runtimeSymbolsByTarget']) {
+    return [...runtimeSymbolsByTarget.entries()].map(([packageName, symbols]) => ({
+        packageName,
+        runtimeSymbols: symbols.size,
+        top: [...symbols.entries()]
+            .sort((a, b) => b[1].size - a[1].size || a[0].localeCompare(b[0]))
+            .slice(0, 6)
+            .map(([symbol, files]) => `${symbol}(${files.size})`),
+    })).sort((a, b) => b.runtimeSymbols - a.runtimeSymbols || a.packageName.localeCompare(b.packageName))
+}
+
+function aggregateTurbulence(rows: readonly TurbulenceRow[]) {
+    const grouped = new Map<string, TurbulenceRow[]>()
+    for (const row of rows) {
+        const existing = grouped.get(row.packageName) ?? []
+        existing.push(row)
+        grouped.set(row.packageName, existing)
+    }
+
+    return [...grouped.entries()].map(([packageName, files]) => {
+        const total = files.reduce((sum, row) => sum + row.turbulence, 0)
+        const maxFile = [...files].sort((a, b) => b.turbulence - a.turbulence || a.file.localeCompare(b.file))[0] ?? null
+        return {packageName, files: files.length, total, average: files.length === 0 ? 0 : total / files.length, maxFile}
+    }).sort((a, b) => b.average - a.average || a.packageName.localeCompare(b.packageName))
+}
+
+function debtRatio(current: number, budget: number, comparison: PressureAxis['comparison']): number {
+    return comparison === 'gte' ? budget / Math.max(1, current) : budget === 0 ? 0 : current / budget
+}
+
+function axisPassed(current: number, budget: number, comparison: PressureAxis['comparison']): boolean {
+    return comparison === 'gte' ? current >= budget : current <= budget
+}
+
+function axis(config: PressureAxisConfig, current: number, worstOffender: string): PressureAxis {
+    return {
+        name: config.name,
+        metricKey: config.metricKey,
+        current,
+        budget: config.budget,
+        comparison: config.comparison,
+        passed: axisPassed(current, config.budget, config.comparison),
+        debtRatio: debtRatio(current, config.budget, config.comparison),
+        worstOffender,
+    }
+}
+
+async function computePressureAxes(): Promise<PressureAxis[]> {
+    const packages = await discoverSystemPackages()
+    const packageNames = packages.map(pkg => pkg.dirName)
+    const graph = await buildSystemGraph(packages)
+    const cognitive = await measureCognitiveComplexity(graph.files)
+    const cyclomatic = await measureCyclomaticComplexity(graph.files)
+    const maintainability = await measureMaintainability(graph.files, cyclomatic)
+    const turbulence = await measureTurbulence(graph.files)
+    const packageTurbulence = aggregateTurbulence(turbulence)
+    const boundaries = measureBoundaries(graph.files, graph.edges, packageNames)
+    const runtimeFanIn = runtimeFanInRows(graph.runtimeSymbolsByTarget)
+    const maxCrap = [...cyclomatic].sort((a, b) => b.crapZeroCoverage - a.crapZeroCoverage)[0]
+
+    return [
+        axis(PRESSURE_AXIS_CONFIGS[0], cognitive[0]?.score ?? 0, cognitive[0] ? `${cognitive[0].file}:${cognitive[0].line} ${cognitive[0].name}` : 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[1], cyclomatic[0]?.score ?? 0, cyclomatic[0] ? `${cyclomatic[0].file}:${cyclomatic[0].line} ${cyclomatic[0].name}` : 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[2], maintainability[0]?.maintainabilityIndex ?? 100, maintainability[0]?.file ?? 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[3], maxCrap?.crapZeroCoverage ?? 0, maxCrap ? `${maxCrap.file}:${maxCrap.line} ${maxCrap.name}` : 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[4], boundaries.boundaryProfiles[0]?.ratio ?? 0, boundaries.boundaryProfiles[0]?.packageName ?? 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[5], boundaries.subdirProfiles[0]?.ratio ?? 0, boundaries.subdirProfiles[0]?.packageName ?? 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[6], boundaries.aggregateBci, boundaries.pairMetrics[0]?.pair ?? 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[7], runtimeFanIn[0]?.runtimeSymbols ?? 0, runtimeFanIn[0]?.packageName ?? 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[8], turbulence[0]?.turbulence ?? 0, turbulence[0]?.file ?? 'n/a'),
+        axis(PRESSURE_AXIS_CONFIGS[9], packageTurbulence[0]?.average ?? 0, packageTurbulence[0]?.packageName ?? 'n/a'),
+    ]
+}
+
+async function readLegacyPressureReport(metricId: string): Promise<LegacyPressureReport | null> {
+    try {
+        return JSON.parse(await readFile(join(REPORTS_DIR, `${metricId}.json`), 'utf8')) as LegacyPressureReport
+    } catch (error) {
+        if (error instanceof Error && 'code' in error && (error as NodeJS.ErrnoException).code === 'ENOENT') return null
+        throw error
+    }
+}
+
+async function compareWithLegacyReports(axes: readonly PressureAxis[]): Promise<string[]> {
+    const mismatches: string[] = []
+    for (const axis of axes) {
+        const config = PRESSURE_AXIS_CONFIGS.find(c => c.name === axis.name)
+        if (!config) continue
+        const legacy = await readLegacyPressureReport(config.metricId)
+        if (!legacy) continue
+        const checks: readonly (readonly [string, unknown, unknown])[] = [
+            ['current', axis.current, legacy.current],
+            ['budget', axis.budget, legacy.budget],
+            ['comparison', axis.comparison, legacy.comparison],
+            ['debtRatio', axis.debtRatio, legacy.details?.debtRatio],
+            ['worstOffender', axis.worstOffender, legacy.details?.worstOffender],
+        ]
+        for (const [field, current, expected] of checks) {
+            if (!Object.is(current, expected)) {
+                mismatches.push(`${axis.name}.${field}: computed=${String(current)} legacy=${String(expected)}`)
+            }
+        }
+    }
+    return mismatches
+}
+
+function computeRscd(axes: readonly PressureAxis[]): {rscd: number; topFiveRatiosForRscd: number[]} {
+    const topFiveRatiosForRscd = axes.map(axis => axis.debtRatio).sort((a, b) => b - a).slice(0, 5)
+    const meanTopFive = topFiveRatiosForRscd.reduce((sum, value) => sum + value, 0) / Math.max(1, topFiveRatiosForRscd.length)
+    return {rscd: (topFiveRatiosForRscd[0] ?? 0) + 0.25 * meanTopFive, topFiveRatiosForRscd}
+}
+
+function failureMessage(axes: readonly PressureAxis[], rscd: number, sanityMismatches: readonly string[]): string {
+    return [
+        `RSCD ${rscd} exceeds target 1.0.`,
+        ...axes.filter(axis => !axis.passed).map(axis => `${axis.name}: current=${axis.current}, budget=${axis.budget}, debtRatio=${axis.debtRatio}, offender=${axis.worstOffender}`),
+        ...sanityMismatches.map(mismatch => `legacy mismatch: ${mismatch}`),
+    ].join('\n')
+}
+
+describe('complexity pressure axes', () => {
+    it('preserves the legacy script pressure rollup semantics', async () => {
+        const axes = await computePressureAxes()
+        const sanityMismatches = await compareWithLegacyReports(axes)
+        const failingAxes = axes.filter(axis => !axis.passed).map(axis => axis.name)
+        const {rscd, topFiveRatiosForRscd} = computeRscd(axes)
+        const message = failureMessage(axes, rscd, sanityMismatches)
+
+        await recordHealthMetric({
+            metricId: 'pressure-axes',
+            metricName: 'Complexity Pressure Axes',
+            description: 'Consolidated 10-axis complexity-pressure rollup',
+            category: 'Complexity',
+            current: rscd,
+            budget: 1.0,
+            comparison: 'lte',
+            unit: 'rscd',
+            details: {axes, failingAxes, topFiveRatiosForRscd, rscd},
+        })
+
+        expect(sanityMismatches, sanityMismatches.join('\n')).toEqual([])
+        for (const pressureAxis of axes) {
+            if (pressureAxis.comparison === 'gte') {
+                expect.soft(pressureAxis.current, message).toBeGreaterThanOrEqual(pressureAxis.budget)
+            } else {
+                expect.soft(pressureAxis.current, message).toBeLessThanOrEqual(pressureAxis.budget)
+            }
+        }
+        expect(rscd, message).toBeLessThanOrEqual(1.0)
+    }, 120000)
+})
